requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0 
sqlalchemy>=2.0.0         
psycopg2-binary>=2.9.0    
alembic>=1.13.0           
python-dotenv>=1.0.0     
google-genai>=0.2.0    